requirements.txt

fastapi
uvicorn
easyocr
pillow
python-multipart
pydantic
